Redirect Already Logged in Users to Maps
If a user was already logged in and tried to go to the homepage, they only had the option to log out, not go to maps, so I just redirected logged in user to maps. Note, we will need to put logout and profile settings buttons on this page now.